Use `run` instead of `shell` so that MD conversions work to names with spaces
Before we used shell, and used shell piping to pipe it to the new file.
Previously just the supplied name was single quoted (this would break if there
were single quotes in the name), and the saved file was not quoted at all,
breaking it when spaces were in the name.

This change also should hopefully work on Windows too, since Windows
doesn't have shell piping.